Refactor server.mjs to modularize IP info and i18n functions, update view data generation for English language
